Move the code into epoch_state module and remove the simulation module
